removed generic parameter from QuickGrid, renamed grid files
