--- src/components/Grid/index.ts
@@ -1,2 +1,0 @@
-export * from './Grid.Props';
-export * from './Grid';
