+++ src/components/QuickGrid/QuickGrid.scss
