+++ src/components/QuickGrid/index.ts
@@ -1,0 +1,2 @@
+export * from './QuickGrid.Props';
+export * from './QuickGrid';
